Add workaround for missing imported css files
Sometimes a css file that is being imported by the preprocess (suitcss)
command is built/fetched by another command like bower.  When doing the
initial dependencies check with '--depend' the command will cause an
error if the file is missing.  This is a workaround for that and shows
a warning that can usually be ignored.  The suitcss command will be
called again when actually building the css and if it causes an error
then the make will/should stop.